requirements.txt
# インストールしたいパッケージを記載
line-bot-sdk
flask
azure-cognitiveservices-vision-face
msrestazure